% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{dataStandardize}
\alias{dataStandardize}
\title{Data standardization}
\usage{
dataStandardize(fpkm, standardize = TRUE)
}
\arguments{
\item{fpkm}{A matrix of gene expression represented as FPKM. Each row represent one gene (named by EnsemblID) and each column represent one sample}
}
\value{
A matrix of standardized FPKM
}
\description{
Standardize FPKM input by log10-transformation followed
by central scaled according to gene expression average 
and sd in the training data. Missing expression values 
are randomly imputated based on the training data average
and sd.
}
