% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictNMI}
\alias{predictNMI}
\title{Neuron Maturity Index (NMI) estimation}
\usage{
predictNMI(fpkm, standardize = TRUE, returnFpkmStd = FALSE)
}
\arguments{
\item{fpkm}{A matrix of gene expression represented by FPKM. Each row represent one gene (named by EnsemblID) and each column represent one sample}

\item{standardize}{To determine whether standardization should be applied to fpkm matrix first}

\item{returnFpkmStd}{To determine whether the standardlized FPKM of samples should be returned.}
}
\value{
A list with at least two components. The first component (modularNMI) is a matrix of modular NMI, with 109 columns each of which shows NMI of one module. The second component (overallNMI) is a matrix of overall NMI, with three columns representing NMI-all, NMI-discriminable and Neuron Activity Index (NAI). The third element is a matrix of standardized FPKM if returnFpkmStd is TRUE.
}
\description{
Taking gene expressions of neuron samples (in FPKM), this
function predicts the Neuron Maturity Index (NMI) which
is an estimate of degree of neuron maturation. The Index
ranges between 0 and 1. The more it is close to one, the
more mature the neuron sample is. The models were trained
using the mature and immature neuron single cell RNA-seq
data in Darmanis, et al. PNAS, 2015.
}
