% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{tpmToFpkm}
\alias{tpmToFpkm}
\title{Convert TPM to FPKM}
\usage{
tpmToFpkm(tpm, geneLength)
}
\arguments{
\item{tpm}{A matrix of gene expression represented as TPM. Each row represent one gene and each column represent one sample}

\item{geneLength}{A vector with the same length as the row number of tpm matrix, representing length of genes in bp}
}
\value{
A matrix of FPKM
}
\description{
Convert TPM to FPKM based on the given gene length. This is
useful when gene expression in neurons are represented as TPM, 
e.g. when quantifying with RSEM.
}
